Move env info helper to shared (oh-tab-xob)
