+++ src/shared/__tests__/environmentInformation.test.ts
@@ -36,4 +36,3 @@ describe('formatEnvironmentInformation', () => {
     expect(text).toContain('- none');
   });
 });
-
